apps/backend/core_backup/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
supabase==2.6.0
pygit2==1.14.1
